+++ src/alerts/mod.rs
@@ -1,3 +1,3 @@
-pub mod drift;
+pub mod base;
 pub mod spc;
 pub mod types;
