+++ src/main.rs
@@ -4,7 +4,7 @@ mod kafka;
 mod rabbitmq;
 mod sql;
 
-use crate::alerts::drift::DriftExecutor;
+use crate::alerts::base::DriftExecutor;
 use crate::api::metrics::metrics_app;
 use crate::api::route::AppState;
 use crate::api::setup::{create_db_pool, setup_logging};
